make repo unpacked and simple to rpkg
We would like to make the repo unpacked in rpkg's terminology, so that
it is easier to view and change source files.

However, seems unpacked repo does not work with individual patch files.
So we apply Macports patches directly to the source files.

Besides, we move spec file to the top level directory, since this is the
default layout.